func/requirements.txt
--extra-index-url https://backend:dc222fa6-452f-4253-a2fd-a52e7ba2e7e9@nexus.sigame.com.br/repository/pypi/simple
motor==3.0.0
flask==2.1.2
heimdall-client==2.3.1
python-decouple==3.6
etria_logger==0.1.10
aioboto3==9.6.0
mepho==0.1.0